+++ src/Resource/EnvelopeTemplateTag.php
@@ -55,6 +55,9 @@ class EnvelopeTemplateTag extends BaseResource
     /** @var int|null */
     public $width;
 
+    /** @var int|null */
+    public $height;
+
     /** @var int */
     public $scale;