+++ CHANGELOG.md
@@ -23,6 +23,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 - Add `User.autoscrollTags`
 - Add `AccountSecurity.discardedEnvelopeRetention`
 - Add `EnvelopeProperties.timestampDocuments`, `EnvelopeProperties.timestampAuditlog`, `EnvelopeProperties.timestampingAuthorities`, `EnvelopeProperties.sendCompleted`
+- Add `AccountSecurity.continuousSigning`
 
 ## [1.10.0]
 ### Added