+++ CHANGELOG.md
@@ -7,7 +7,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 ## [Unreleased]
 - Add new fields for IdentityProvider
 - Add new field $userId for User
-
+- Add new field $signatureValidity for EnvelopeDocument
 ## [1.11.0]
 ### Added
 - Add new oidc fields for `AccountSecurity` resource