+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 - Add new fields for IdentityProvider
 - Add new field $userId for User
 - Add new field $signatureValidity for EnvelopeDocument
+- Add new field $height for EnvelopeTag and EnvelopeTemplateTag
 - Add new fields $identificationNumber, $address for `EnvelopeRecipient` and `EnvelopeTemplateRecipient`
 
 ## [1.11.0]